LanUdevAuto: moving functionality into lib
