+++ src/lib/network/network_autoyast.rb
@@ -115,7 +115,29 @@ module Yast
       NetworkService.EnableDisableNow
     end
 
-  private
+    private
+    # Checks if given profile contains device names in old style format
+    #
+    # OLd style device name format is eth-id-<mac> or eth-bus-<busid>.
+    # These names has to be converted into new-style (eth0) and name persistency
+    # is enabled using udev rule which is generated later.
+    def oldStyle(ay)
+      ay = deep_copy(ay)
+      old_style_found = false
+      Builtins.foreach(Ops.get_list(ay, "interfaces", [])) do |interface|
+        if Builtins.issubstring(Ops.get_string(interface, "device", ""), "-id-")
+          old_style_found = true
+        end
+        if Builtins.issubstring(
+          Ops.get_string(interface, "device", ""),
+          "-bus-"
+          )
+          old_style_found = true
+        end
+      end
+      Builtins.y2milestone("old-style found:%1", old_style_found)
+      old_style_found
+    end
 
     # Merges two devices map into one.
     #
